Fold ailetos_rt_glue into lib.rs and delete the module
The module no longer justified its existence: start_executor, start_ctrlc_handler,
start_notification_watcher, and make_env are all small enough to live directly
in lib.rs alongside DagShell. JoinWaiter moves there too.
